Split Algorithms
