--- Planner/include/algorithms.h
@@ -1,158 +1,0 @@
-#include "state_wrapper.h"
-#include <unordered_map>
-#include <queue>
-
-#ifndef SEARCH_DEF
-#define SEARCH_DEF
-
-using namespace std;
-
-template<typename X>
-class Comparator {
-public:
-    int operator()(X *p_1, X *p_2) const {
-        return *p_1 > *p_2;
-    }
-};
-
-template<typename X>
-X *Astar(X *initial, X *goal, long long &examined, long long &mem) {
-    priority_queue<X *, vector<X *>, Comparator<X> > agenda;
-    unordered_map<unsigned long long, bool> closed;
-    agenda.push(initial);
-    examined = 0;
-    mem = 1;
-    while (agenda.size() > 0) {
-        if ((long long) (agenda.size() + closed.size()) > mem)
-            mem = agenda.size() + closed.size();
-        examined++;
-        X *s = agenda.top();
-        agenda.pop();
-        if (*s >= *goal) {
-            return s;
-        }
-        closed[s->getHash()] = true;
-        vector<X *> children = s->expand();
-        for (unsigned int i = 0; i < children.size(); i++)
-            if (!closed[children.at(i)->getHash()]) {
-                children.at(i)->SetHvalue(children.at(i)->GetDepth() + children.at(i)->estimate());
-                agenda.push(children.at(i));
-            }
-    }
-    return nullptr;
-}
-
-template<typename X>
-X *BFS(X *initial, X *goal, long long &examined, long long &mem) {
-    priority_queue<X *, vector<X *>, Comparator<X> > agenda;
-    unordered_map<string, bool> closed;
-    agenda.push(initial);
-    examined = 0;
-    mem = 1;
-    while (agenda.size() > 0) {
-        if ((long long) (agenda.size() + closed.size()) > mem)
-            mem = agenda.size() + closed.size();
-        examined++;
-        X *s = agenda.top();
-        agenda.pop();
-        if (*s >= *goal) {
-            cout << "Closed size: " << closed.size() << endl;
-            return s;
-        }
-        closed[s->getId()] = true;
-        vector<X *> children = s->expand();
-        for (unsigned int i = 0; i < children.size(); i++)
-            if (!closed[children.at(i)->getId()]) {
-                children.at(i)->SetHvalue(children.at(i)->estimate());
-                agenda.push(children.at(i));
-            }
-    }
-    return nullptr;
-}
-
-template<typename X>
-X *IDAstar(X *initial, X *goal, long long &examined, long long &mem) {
-    double current_cost;
-    double cost = 0;
-
-    initial->SetHvalue(initial->estimate());
-
-    while (true) {
-
-        queue<X *> agenda = queue<X *>();
-        unordered_map<X, bool> closed = unordered_map<X, bool>();
-
-        agenda.push(initial);
-        examined = 0;
-        mem = 1;
-        cout << "Examined: " << examined << endl;
-        cout << "Mem: " << mem << endl;
-
-        bool continue_search = true;
-        while (continue_search && agenda.size() > 0) {
-            cout << "Examined: " << examined << endl;
-            cout << "Mem: " << mem << endl;
-            if ((long long) (agenda.size() + closed.size()) > mem)
-                mem = agenda.size() + closed.size();
-            examined++;
-
-            X *s = agenda.front();
-            agenda.pop();
-
-            current_cost = s->GetHvalue();
-
-            if (*s >= *goal) {
-                cout << "Closed size: " << closed.size() << endl;
-                return s;
-            }
-            closed[*s] = true;
-
-
-            if (current_cost > cost) { // Check if h value is larger than the current cost
-                cost = current_cost;
-                continue_search = false;
-            } else { // Else add the expanded states in the agenda queue
-                vector<X *> children = s->expand();
-                for (unsigned int i = 0; i < children.size(); i++)
-                    if (!closed[*children.at(i)]) {
-                        children.at(i)->SetHvalue(children.at(i)->estimate());
-                        agenda.push(children.at(i));
-                    }
-            }
-        }
-        cout << "Examined: " << examined << endl;
-        cout << "Mem: " << mem << endl;
-
-        // Examined all possible states and none matched the goal
-        if (continue_search && agenda.size() <= 0) return nullptr;
-    }
-}
-
-template<typename X>
-X *DFS(X *initial, X *goal, long long &examined, long long &mem) {
-    stack<X *> frontier;
-    unordered_map<X, bool> closed;
-    frontier.push(initial);
-    examined = 0;
-    mem = 1;
-    while (frontier.size() > 0) {
-        if ((long long) (frontier.size() + closed.size()) > mem)
-            mem = frontier.size() + closed.size();
-        examined++;
-        X *s = frontier.top();
-        frontier.pop();
-        if (*s >= *goal) {
-            cout << "Closed size: " << closed.size() << endl;
-            return s;
-        }
-        closed[*s] = true;
-        vector<X *> children = s->expand();
-        for (unsigned int i = 0; i < children.size(); i++)
-            if (!closed[*children.at(i)]) {
-                frontier.push(children.at(i));
-            }
-    }
-    return nullptr;
-}
-
-#endif
